+++ tree-sitter-hird/queries/highlights.scm
@@ -151,6 +151,7 @@
 
 [
   "type"
+  "alias"
   "effect"
   "tool"
   "actor"